[feat] Names 클래스 구현
Names 입력받은 이름을 검증/포장하는 클래스 구현
ALL 코드 컨벤션 및 구조 변경에 따른 리팩토링
