+++ src/main/java/domain/GameResult.java
@@ -7,7 +7,6 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 public class GameResult {
     private final Map<String, Boolean> playerResult = new HashMap<>();
